requirements.txt
asgiref==3.7.2
boto3==1.35.53
botocore==1.35.53
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
google_search_results==2.4.2
idna==3.7
jmespath==1.0.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.3
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
